Phase 1 update
I've been working on COMS for a few days now. Sadly, this is the first commit I've performed on this rep.

At this point, I have the phase 1 layout complete, the 2 player buttons are interactable & the counters for each work. The "finalize" button also takes the user to the 2nd activity.

The problems right now is that the winner data isn't being sent/ received to/by the 2nd activity. I'm sure it has something to do with how I've implemented my "startActivity()".