Fixed intent data send/ receive
The problem was that the Intent didn't like the data being sent as an int type; it wanted a string. So I casted the data as strings, and now it sends & displays the data!

Next, I need to edit it so that it also displays which player had the higher score/ won.