Final Layout Decision
I've tried to implement a dynamic layout: First the app asks how many users will play, then asks for that many names, then displays that many TV's & BTN's. This is challengning, and with the time restrictions I'm placing on myself, I've decided to program the interface to always allow for 5 players.

This allows the users to choose up to 5 players, or at least 2. In the furture I will return to implementing a dynamic layout.