Soved Getting Largest Scores Index
After searching online for help, I found an easy solution: Search through the arraylist for the max value, while using a counter to track the index. Then, return the index.

App now allows for 5 players. The highest score is sent & displayed with the respective player when the finalize/ scores button is pressed.

One issue that remains is to now program it so that when multiple players have the same highest score, all of them are displayed. i.e., if player 1 & player 3 have the high score of 17, only player 1 is currently displayed. Both players need to be displayed as the winners.